actuall pass the config
